refactor(routes): Abstract route middleware, move middleware logic to seperate fils
